+++ components/organizer/organisms/Lessons.vue
@@ -3,9 +3,12 @@
     <div class="subtitle-1 font-weight-bold">
       <div>指導者/生徒を募集する</div>
       <div class="text-center px-4 mt-2">
-        <router-link to="/organizer/lessons/new" class="lesson-link">
+        <router-link to="/lessons/new" class="lesson-link" target="_blank">
           <v-btn color="primary" outlined block>
             募集情報入力フォームへ
+            <v-icon small class="ml-1">
+              mdi-open-in-new
+            </v-icon>
           </v-btn>
         </router-link>
       </div>
