+++ layouts/default.vue
@@ -26,7 +26,7 @@
         <img border="0" width="300" height="250" alt="" src="https://www26.a8.net/svt/bgt?aid=220121080377&wid=001&eno=01&mid=s00000021147002010000&mc=1"></a>
         <img border="0" width="1" height="1" src="https://www19.a8.net/0.gif?a8mat=3N1YIG+68GFUA+4J66+BYT9D" alt="">
       </div>
-      <div class="footer-links">
+      <div v-if="!!$route.path.indexOf('/organizer')" class="footer-links">
         <div v-for="link in filteredFooterLinks()" :key="link.title" class="link white--text">
           <router-link :to="link.to">
             {{ link.title }}
@@ -108,15 +108,15 @@ export default {
           ...this.footerLinks,
           {
             title: 'チームを登録する',
-            to: '/organizer/teams/new'
+            to: '/teams/new'
           },
           {
             title: 'オンラインサロンを申請する',
-            to: '/organizer/salons/new'
+            to: '/salons/new'
           },
           {
             title: 'クラウドファンディングを申請する',
-            to: '/organizer/funds/new'
+            to: '/funds/new'
           }
         ]
       } else {
